Remove xowiki dependency from CKEditor4
Rationale: CKEditor4 is EOL for its free version and the paid LTS version will be discontinued in December 2026.

We now do not set an explicit default for the preferred richtext editor. Instead, the first richtext-* package that will be installed with check whether a preferred editor was chosen. If not, they will set themselves as the xowiki editor. This should not change the behavior for existing installations already using ckeditor4, or a different editor.